// src/app/modules/estudiante/models/estudianteperfil.model.ts
export interface EstudiantePerfil {
    id_usuario: number;
    nombre_usuario: string;
    contrasena_usuario: string;
    rol_usuario: number;
    id_estudiante: number;    
    nombres_estudiante: string;
    apellidos_estudiante: string;
    cedula_estudiante: string;
    fechaNacimiento_estudiante: string;
    edad_estudiante: number;
    direccion_estudiante: string;
    telefono_estudiante: string;
    email_estudiante: string;
    nivelEducacion_estudiante: string;
    promedioAnterior_estudiante: number;
    medio_estudiante: number;
    estado_estudiante: string;
  }